+++ lib/phx_live_storybook/live/entry/component_iframe_live.ex
@@ -10,7 +10,8 @@ defmodule PhxLiveStorybook.ComponentIframeLive do
     {:ok,
      assign(socket,
        otp_app: session["otp_app"],
-       backend_module: session["backend_module"]
+       backend_module: session["backend_module"],
+       assets_path: session["assets_path"]
      ), layout: {PhxLiveStorybook.LayoutView, "live_iframe.html"}}
   end